Renamed a function (nfc).
`GetTestFullFromLine` returns a test name as a `Case.Test` string. Its
result is used as a filter expression with minimal processing (splitting
at the dot to extract some case and test). Rename it to
`GetTestFilterFromLine` to make clear we return a filter expression. In
a later commit we will change it to return filters for parameterized
tests as well.