Made it possible to select parameterized tests.
Parameterized tests lead to different test names. Instead of simply
concatenating case and test both a pre- and a post fix are added.

Adjust `GetTestFilterFromLine` so it inserts wildcards into the filter
where appropriate. This makes it possible to use e.g.,
`GTestRunUnderCursor` to run a parameterized test; with this change it
is still not possible to select a parameterized case (this is harder
since a case might contain both parameterized and unparameterized tests
and googletest filter expressions cannot contain alternatives like
`(\*/|.?)`.

This fixes #14.