Narrower filter for parameterized tests (#19)
The handling of parameterized tests introduced in `7fb016c794f` was not
narrow enough. With a parameterized test `Foo.Bar` under the under the
cursor it would not only match e.g., `0/Foo.Bar/0` but also
`0/LeFoo.Bar/0` or `0/Foo.BarBaz/0`.

With this patch we explicitly include the slashes in the generated
filter expression (e.g., filter for `*/Foo.Bar/*` instead of
`*Foo.Bar*`).